+++ shared/js/background/classes/company.es6.js
