+++ shared/js/background/classes/score.es6.js
@@ -1,5 +1,5 @@
-const tosdr = require('../data/tosdr')
-const constants = require('../data/constants')
+const tosdr = require('../../../data/tosdr')
+const constants = require('../../../data/constants')
 const tosdrRegexList = Object.keys(tosdr).map(x => new RegExp(x))
 const tosdrClassMap = {'A': -1, 'B': 0, 'C': 0, 'D': 1, 'E': 2} // map tosdr class rankings to increase/decrease in grade
 const siteScores = ['A', 'B', 'C', 'D']
