+++ shared/js/background/debug.es6.js
@@ -2,7 +2,7 @@
  * This exposes some modules we use for testing via the background page console.
  * NOTE this is not added to the release version of the extension
  */
-const settings = require('./settings')
+const settings = require('./settings.es6')
 const abp = require('abp-filter-parser')
 
 window.dbg = {
