+++ shared/js/background/load.es6.js
