+++ shared/js/background/surrogates.es6.js
