+++ shared/js/background/utils.es6.js
