+++ test/requireHelper.js
@@ -1,13 +1,13 @@
 window.bkg = {
-    settings: require('../shared/js/settings'),
-    ATB: require('../shared/js/atb'),
-    Companies: require('../shared/js/companies'),
-    tabManager: require('../shared/js/tabManager'),
-    https: require('../shared/js/https'),
-    trackers: require('../shared/js/trackers'),
-    utils: require('../shared/js/utils'),
-    surrogates: require('../shared/js/surrogates'),
-    Site: require('../shared/js/site'),
+    settings: require('../shared/js/background/settings.es6'),
+    ATB: require('../shared/js/background/atb.es6'),
+    Companies: require('../shared/js/background/companies.es6'),
+    tabManager: require('../shared/js/background/tab-manager.es6'),
+    https: require('../shared/js/background/https.es6'),
+    trackers: require('../shared/js/background/trackers.es6'),
+    utils: require('../shared/js/background/utils.es6'),
+    surrogates: require('../shared/js/background/surrogates.es6'),
+    Site: require('../shared/js/background/classes/site.es6'),
     abp: require('abp-filter-parser'),
     tosdr: require('../shared/data/tosdr')
 }
